requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
motor==3.5.1
redis==5.0.8
numpy==2.1.1
